requirements.txt
pydicom
matplotlib
xml-python
numpy
opencv-python
xlrd
Pillow
pandas
scipy
tqdm
qqdm
openpyxl==3.0.10
torchmetrics
imutils
kaleido
albumentations
plotly
